fix -Wsign-compare

+++ test/timing-decoder.cpp
@@ -41,7 +41,7 @@ int main(int argc, char const *argv[]) {
 
     /* fill up our data-buffer with single "bytes", formed from the 4-byte
      * numbers returned by rd() */
-    while (data.size() < numberOfBytesToProcess) {
+    while ((long long int)data.size() < numberOfBytesToProcess) {
         unsigned int number = rd();
         data.push_back(number);
         data.push_back(number >> 8);
